Fix tiiiiiiny inconsistency vs osu-stable resulting in one extra slider tick on some sliders.

+++ osu.Game.Modes.Taiko/Objects/DrumRoll.cs
@@ -73,7 +73,7 @@ namespace osu.Game.Modes.Taiko.Objects
                 if (TickTimeDistance == 0) yield break;
 
                 bool first = true;
-                for (double t = StartTime; t < EndTime + TickTimeDistance; t += TickTimeDistance)
+                for (double t = StartTime; t < EndTime + (int)TickTimeDistance; t += TickTimeDistance)
                 {
                     yield return new DrumRollTick()
                     {
